+++ wemake_python_styleguide/presets/topics/complexity.py
@@ -4,6 +4,7 @@ from typing_extensions import Final
 
 from wemake_python_styleguide.visitors.ast.complexity import (
     access,
+    calls,
     classes,
     counts,
     function,
@@ -38,4 +39,6 @@ PRESET: Final = (
     overuses.ExpressionOveruseVisitor,
 
     access.AccessVisitor,
+
+    calls.CallChainsVisitor,
 )